SAK-6361 Removed check for hiding the checkbox.
hiddenLock is not referenced anywhere in PermissionHelperAction so will always be empty.
editViewRoster is never setup in PermissionHelperAction so will always be empty.
The permissions helper never gets called for the site prefix so I'm dropping the check on site.vireRoster as well.

The hidden checkbox would also have broken realms where the permission came in through an additional realm.

git-svn-id: https://source.sakaiproject.org/svn/authz/branches/SAK-6361@123027 66ffb92e-73f9-0310-93c1-f5514f145a0a